==> scripts/1 Thesis - ADReSS - Data Preparation.ipynb <==
# ADReSS Dataset

By Jimuel Celeste, Jr. 

Objective: To prepare a clean metadata for the ADReSS Dataset.


import os 
import pandas as pd

## Train

train1 = '/Users/jimuelcelestejr/Downloads/Dataset/ADReSS/archive/ADReSS-IS2020-data_train/cd_meta_data_train.txt'
train1_df = pd.read_csv(train1, sep=';')
train1_df['dx'] = 'AD'
train1_df.head()

train2 = '/Users/jimuelcelestejr/Downloads/Dataset/ADReSS/archive/ADReSS-IS2020-data_train/cc_meta_data_train.txt'
train2_df = pd.read_csv(train2, sep=';')
train2_df['dx'] = 'NC'
train2_df.head()

train_df = pd.concat([train1_df, train2_df])
train_df

train_df.shape, train1_df.shape[0] + train2_df.shape[0]

train_df.columns

for column in train_df.columns:
    train_df = train_df.rename(columns={column: column.strip()})
train_df.columns

for column in train_df.columns: 
    train_df[column] = train_df[column].apply(lambda x: x.strip() if type(x) == str else x)

train_df = train_df.rename(columns={
    'ID': 'record_id', 
    'gender': 'sex'
})
train_df['dx_binary'] = train_df['dx'].apply(lambda x: 1 if x == 'AD' else 0)
train_df['subject_id'] = train_df['record_id']
train_df['filename'] = train_df['record_id'].apply(lambda x: x + '.wav')
train_df['sex'] = train_df['sex'].apply(lambda x: 'M' if x == 'male' else 'F')
train_df = train_df[[
    'filename', 
    'record_id', 
    'subject_id', 
    'age', 
    'sex', 
    'mmse', 
    'dx', 
    'dx_binary'
]].reset_index(drop=True)
train_df

## Test

test = '/Users/jimuelcelestejr/Downloads/Dataset/ADReSS/archive/ADReSS-IS2020-data 2_test/ADReSS-meta_data_test.txt'
test_df = pd.read_csv(test, sep=';')
test_df.head()

test_df.columns

for column in test_df.columns:
    test_df = test_df.rename(columns={column: column.strip()})
test_df.columns

for column in test_df.columns: 
    test_df[column] = test_df[column].apply(lambda x: x.strip() if type(x) == str else x)

test_metadata = '/Users/jimuelcelestejr/Downloads/Dataset/ADReSS/archive/ADReSS-IS2020-data 2_test/meta_data_test.txt'
test_metadata_df = pd.read_csv(test_metadata, sep=';')
test_metadata_df.head()

test_metadata_df.columns

for column in test_metadata_df.columns:
    test_metadata_df = test_metadata_df.rename(columns={column: column.strip()})
test_metadata_df.columns

for column in test_metadata_df.columns: 
    test_metadata_df[column] = test_metadata_df[column].apply(lambda x: x.strip() if type(x) == str else x)

set(test_df['ID']) - set(test_metadata_df['ID'])

set(test_metadata_df['ID']) - set(test_df['ID'])

set(test_metadata_df['ID']) == set(test_df['ID'])

Same; I'll only use test_df.

test_df.head()

test_df = test_df.rename(columns={
    'ID': 'record_id', 
    'gender': 'sex', 
    'Label': 'dx_binary'
})
test_df['dx'] = test_df['dx_binary'].apply(lambda x: 'AD' if x == 1 else 'NC')
test_df['subject_id'] = test_df['record_id']
test_df['filename'] = test_df['record_id'].apply(lambda x: x + '.wav')
test_df['sex'] = test_df['sex'].apply(lambda x: 'F' if x == 1 else 'M')
test_df = test_df[[
    'filename', 
    'record_id', 
    'subject_id', 
    'age', 
    'sex', 
    'mmse', 
    'dx', 
    'dx_binary'
]].reset_index(drop=True)
test_df.head()

## Merge: train + test

merged_df = pd.concat([train_df, test_df])
merged_df = merged_df.sort_values(by='filename')
merged_df = merged_df.reset_index(drop=True)
merged_df

train_df.shape[0] + test_df.shape[0]

merged_df[merged_df.duplicated()]

merged_df['dx'].value_counts()

## Save: metadata.csv

output_dir = '/Users/jimuelcelestejr/Downloads/Dataset/ADReSS'
output_file = os.path.join(output_dir, 'metadata.csv')
merged_df.to_csv(output_file, index=False)

## Check if all files are in the folder

metadata = pd.read_csv(output_file)
i = 0
for file in metadata['filename']:
    file = os.path.join(output_dir)
    if os.path.exists(file):
        i += 1
        # print(f"'{file}' exists.")
    else:
        print(f"'{file}' does not exist.")
print(metadata.shape, i)

## Check if all files in the folder are in the metadata¶

i = 0
for root, dirs, files in os.walk(output_dir):
    for file in files:
        base, ext = os.path.splitext(file)
        if ext == '.wav':
            if file in metadata['filename'].values:
                i += 1
            else:
                print(f"'{file}' not in metadata.")
print(i)
