==> scripts/3 Thesis - ADReSS-M - Data Preparation.ipynb <==
# ADReSS-M Dataset

By Jimuel Celeste, Jr. 

Objective: To prepare a clean metadata for the ADReSS-M Dataset.


import os 
import pandas as pd

## Train

train = '/Users/jimuelcelestejr/Downloads/Dataset/ADReSS-M/archive/ADReSS-M-train/training-groundtruth.csv'
train_df = pd.read_csv(train)
train_df

## Test

test = '/Users/jimuelcelestejr/Downloads/Dataset/ADReSS-M/archive/test-gr/ADReSS-M-test-gr-groundtruth.csv'
test_df = pd.read_csv(test)
test_df = test_df.rename(columns={'addressfname': 'adressfname'})
test_df.head()

test_metadata = '/Users/jimuelcelestejr/Downloads/Dataset/ADReSS-M/archive/test-gr/ADReSS-M-meta.csv'
test_metadata_df = pd.read_csv(test_metadata)
test_metadata_df = test_metadata_df.rename(columns={'addressfname': 'adressfname'})
test_metadata_df.head()

Age, gender, and education are already included in test_df; I will use that.

test_df.shape, test_metadata_df.shape

## Sample

sample = '/Users/jimuelcelestejr/Downloads/Dataset/ADReSS-M/archive/ADReSS-M-sample-gr/sample-gr-groundtruth.csv'
sample_df = pd.read_csv(sample)
sample_df = sample_df.rename(columns={'addressfname': 'adressfname'})
sample_df

## Merge: train + test + sample

merged_df = pd.concat([train_df, test_df, sample_df])
merged_df.head()

merged_df[merged_df['educ'].isna()]

merged_df.shape, train_df.shape[0] + test_df.shape[0] + sample_df.shape[0]

Records without education: 21.

merged_df = merged_df.rename(columns={
    'adressfname': 'record_id', 
    'gender': 'sex'
})
merged_df['filename'] = merged_df['record_id'].apply(lambda x: x + '.wav' if x[0:4] == 'madr' else x + '.mp3')
merged_df['subject_id'] = merged_df['record_id']
merged_df['dx'] = merged_df['dx'].apply(lambda x: 'NC' if x == 'Control' else 'Probable AD')
merged_df['dx_binary'] = merged_df['dx'].apply(lambda x: 1 if x == 'Probable AD' else 0)
merged_df['sex'] = merged_df['sex'].apply(lambda x: 'F' if x.lower() == 'female' else 'M')
merged_df = merged_df[[
    'filename', 
    'record_id', 
    'subject_id', 
    'age', 
    'sex', 
    'educ',
    'mmse',
    'dx', 
    'dx_binary'
]].reset_index(drop=True)
merged_df = merged_df.sort_values(by='filename')
merged_df

merged_df[merged_df.duplicated()]

merged_df['dx'].value_counts()

## Save: metadata.csv

output_dir = '/Users/jimuelcelestejr/Downloads/Dataset/ADReSS-M'
output_file = os.path.join(output_dir, 'metadata.csv')
merged_df.to_csv(output_file, index=False)

## Check if all files are in the folder

metadata = pd.read_csv(output_file)
i = 0
for file in metadata['filename']:
    file = os.path.join(output_dir)
    if os.path.exists(file):
        i += 1
        # print(f"'{file}' exists.")
    else:
        print(f"'{file}' does not exist.")
print(metadata.shape, i)

## Check if all files in the folder are in the metadata

i = 0
for root, dirs, files in os.walk(output_dir):
    for file in files:
        base, ext = os.path.splitext(file)
        if ext in ['.wav', '.mp3']:
            if file in metadata['filename'].values:
                i += 1
            else:
                print(f"'{file}' not in metadata.")
    break # in the root directory only
print(i)
