==> scripts/7 Thesis - Transcript Preparation.ipynb <==
# Transcript Preparation

By: Jimuel Celeste, Jr. 

Objective: To prepare a clean dataset for the transcripts.

import os 

import json
import pandas as pd

## Check: completeness

metadata = '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Combined/metadata.csv'
metadata_df = pd.read_csv(metadata)
metadata_df.head()

input_dir = '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Whisper Transcripts'
i = 0
for file in metadata_df['filename'].values:
    base, ext = os.path.splitext(file)
    json_file = base + '.json'
    json_file = os.path.join(input_dir, json_file)
    
    if os.path.exists(json_file):
        i +=1 
    else:
        print('File does not exist:', json_file)
print(i)

Complete dataset

## Metadata

# new_metadata
def format_ext(file, new_ext):
    base, ext = os.path.splitext(file)
    json_file = base + new_ext
    return json_file 
    
metadata_df['filename'].apply(lambda x: format_ext(x, '.json')).head()

new_metadata_df = metadata_df.copy(deep=True)
new_metadata_df.head()

new_metadata_df['filename'] = new_metadata_df['filename'].apply(lambda x: format_ext(x, '.json'))
new_metadata_df.head()

new_metadata_csv = os.path.join(input_dir, 'metadata.csv')
new_metadata_df.to_csv(new_metadata_csv, index=False)

## Generate text metadata

text_metadata_df = metadata_df.copy(deep=True)
text_metadata_df['filename'] = text_metadata_df['filename'].apply(lambda x: format_ext(x, '.txt'))
text_metadata_df.head()

## Generate text files

def load_json(file):
    with open(file, 'r') as f:
        return json.load(f)

sample = '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Whisper Transcripts/taukdial-169-3.json'
sample_json = load_json(sample)
sample_json

sample_json.keys()

sample_json['text']

sample_json['language']

sample_json['segments']

text_transcripts = {}
languages = {}

for file in new_metadata_df['filename']:
    input_file = os.path.join(input_dir, file)
    data = load_json(input_file)

    text_filename = format_ext(file, '.txt')
    
    text = data['text']
    text_transcripts[text_filename] = text
    
    language = data['language']
    languages[text_filename] = language

languages_df = pd.DataFrame.from_dict(languages, orient='index').reset_index()
languages_df = languages_df.rename(columns={'index': 'filename', 0: 'language'})
languages_df

text_metadata_df = text_metadata_df.merge(languages_df, on='filename', how='left')
text_metadata_df.head()

text_metadata_df['language'].value_counts()

- 708 English 
- 261 Chinese 
- 54 Greek

output_dir = '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Text Transcripts'
text_metadata_csv = os.path.join(output_dir, 'metadata.csv')
text_metadata_df.to_csv(text_metadata_csv, index=False)

## Save text files

def save_text(text, file_path):
    with open(file_path, 'w') as f:
        f.write(text)
    return None
    
for file in text_transcripts:
    file_path = os.path.join(output_dir, file) 
    text = text_transcripts[file]
    save_text(text, file_path)

## Check for completeness

input_dir = '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Text Transcripts'
i = 0
for file in text_metadata_df['filename'].values:
    text_file = os.path.join(input_dir, file)
    
    if os.path.exists(text_file):
        i +=1 
    else:
        print('File does not exist:', textfile)
print(i)

Complete list
