==> scripts/8 Thesis - Splitting - Data Preparation.ipynb <==
# Data Preparation - Splitting

By: Jimuel Celeste, Jr. 

Objective: To split the combined dataset into two for internal cross-validation and external validation.

import os 

import pandas as pd

feature_sets = {
    'eGeMAPS': '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Acoustic Features/egemaps_functionals', 
    'BoAW ComParE 2016': '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Acoustic Features/boaw_size_200_compare_2016_lld', 
    'LFTK': '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Linguistic Features/lftk_features',
    'LFTK Syntax': '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Linguistic Features/syntax_features',
    'LFTK Discourse': '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Linguistic Features/discourse_features',
    'LFTK Lexico-Semantics': '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Linguistic Features/lexico_semantics_features',
    'LFTK Surface': '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Linguistic Features/surface_features',
    'BERT': '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - BERT Features', 
    'wav2vec': '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - wav2vec Features'
}

# check if metadata exists in the folds
for feature_set in feature_sets:
    metadata = os.path.join(feature_sets[feature_set], 'metadata.csv')
    if os.path.exists(metadata):
        print(feature_set, 'ok')
    else:
        print(feature_set, 'does not exist')

## External Validation Sets

metadata = os.path.join(feature_sets['LFTK'], 'metadata.csv') # reference feature set 
metadata_df = pd.read_csv(metadata)
metadata_df.head()

metadata_df['language'].value_counts() # languages

### Chinese

chinese = metadata_df[metadata_df['language']=='zh']
chinese.shape

chinese.head()

chinese['dataset'].value_counts()

### Greek

greek = metadata_df[metadata_df['language']=='el']
greek.shape

greek.head()

greek['dataset'].value_counts()

### English

# English & TAUKADIAL 
english = metadata_df[(metadata_df['language']=='en') & (metadata_df['dataset']=='TAUKADIAL')]
english.size

english.head()

english['dataset'].value_counts()

# control case: english, cookie theft
def get_task_id(record_id):
    return record_id.split('-')[2]

english_cookie_theft = english[english['record_id'].apply(lambda x: get_task_id(x)=='3')]
english_cookie_theft.shape

english_cookie_theft.head()

english_cookie_theft['record_id'].apply(lambda x: get_task_id(x)).value_counts()

Task ids are all 3.

english_others = english[~english['record_id'].isin(english_cookie_theft['record_id'].values)]
english_others.size

english_others.head()

english_others['record_id'].apply(lambda x: get_task_id(x)).value_counts()

### Save: metadata files

Only tasks 1 and 2 are here.

# Saving validation sets in each of the feature set directories

external_validation_sets = {
    'metadata_ev_english_control': english_cookie_theft,
    'metadata_ev_english': english_others,
    'metadata_ev_chinese': chinese,
    'metadata_ev_greek': greek
}

for fs in feature_sets:
    current_dir = feature_sets[fs]
    print(fs)
    for evs in external_validation_sets:
        current_metadata_csv = os.path.join(current_dir, evs + '.csv')
        current_metadata = external_validation_sets[evs]

        current_metadata.to_csv(current_metadata_csv, index=False)
        
        if os.path.exists(current_metadata_csv):
            print('\t +', evs)
        else:
            print('\t -', evs, )


## Internal Cross-Validation

external_cv_record_ids = []
for evs in external_validation_sets:
    current_metadata = external_validation_sets[evs]
    external_cv_record_ids.extend(current_metadata['record_id'].values)
len(external_cv_record_ids)  

metadata_df.shape, metadata_df.shape[0] - 561

462 samples left for internal CV.

internal_cv = metadata_df[~metadata_df['record_id'].isin(external_cv_record_ids)]
internal_cv.shape

internal_cv['language'].value_counts()

internal_cv['dataset'].value_counts()

# Saving internal cross-validation set in each of the feature set directories

for fs in feature_sets:
    print(fs)
    current_dir = feature_sets[fs]
    current_metadata_csv = os.path.join(current_dir, 'metadata_cv_english.csv')
    internal_cv.to_csv(current_metadata_csv, index=False)
    
    if os.path.exists(current_metadata_csv):
        print('\t +', 'metadata_cv_english.csv')
    else:
        print('\t -', 'metadata_cv_english.csv')
