==> scripts/15 Thesis - Experimet Datasets (additional analysis).ipynb <==
# Thesis - Experimet Datasets (additional analysis)

By: Jimuel Celeste, Jr. 

Objective: To describe the distribution of labels by set. 

import os 

import pandas as pd

## CV

metadata = '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Acoustic Features/egemaps_functionals/metadata_cv_english.csv'
df = pd.read_csv(metadata)
df.head()

df.shape

df['dx_binary'].value_counts()

df['dx_binary'].value_counts() / df.shape[0] * 100

df['mmse'].describe()

df[df['dx_binary']==1]['mmse'].describe()

df[df['dx_binary']==0]['mmse'].describe()

df['mmse'].isna().sum()

## EV 1: English 1

metadata = '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Acoustic Features/egemaps_functionals/metadata_ev_english_control.csv'
df = pd.read_csv(metadata)
df.head()

df.shape

df['dx'].value_counts()

df['dx'].value_counts() / df.shape[0] * 100

df['mmse'].describe()

df[df['dx_binary']==1]['mmse'].describe()

df[df['dx_binary']==0]['mmse'].describe()

df['mmse'].isna().sum()

## EV 2: English 2

metadata = '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Acoustic Features/egemaps_functionals/metadata_ev_english.csv'
df = pd.read_csv(metadata)
df.head()

df.shape

df['dx'].value_counts()

df['dx'].value_counts() / df.shape[0] * 100

df['mmse'].describe()

df[df['dx_binary']==1]['mmse'].describe()

df[df['dx_binary']==0]['mmse'].describe()

df['mmse'].isna().sum()

## EV 3: Greek

metadata = '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Acoustic Features/egemaps_functionals/metadata_ev_greek.csv'
df = pd.read_csv(metadata)
df.head()

df.shape

df['dx'].value_counts()

df['dx'].value_counts() / df.shape[0] * 100

df['mmse'].describe()

df[df['dx_binary']==1]['mmse'].describe()

df[df['dx_binary']==0]['mmse'].describe()

## EV 4: Chinese

metadata = '/Users/jimuelcelestejr/Downloads/Dataset/Thesis - Features/Thesis - Acoustic Features/egemaps_functionals/metadata_ev_chinese.csv'
df = pd.read_csv(metadata)
df.head()

df.shape

df['dx'].value_counts()

df['dx'].value_counts() / df.shape[0] * 100

df['mmse'].describe()

df[df['dx_binary']==1]['mmse'].describe()

df[df['dx_binary']==0]['mmse'].describe()

df['mmse'].isna().sum()
